wsputil: Add 'simplified' text header iterator
In large part this is inspired by RFC 2616, however we use a much
simplified syntax that does not require a full blown lexer + parser
setup.

One header + parameters per line.  If the header value is not a token
(matching RFC 2616 / WSP definition of a token) then the value must be
quoted.  All header keys and parameter keys must be tokens.